+++ lib/component.ex
@@ -2357,6 +2357,11 @@ defmodule PrimerLive.Component do
 
   attr(:class, :string, default: nil, doc: "Additional classname.")
 
+  attr(:is_wrap, :boolean,
+    default: false,
+    doc: "Allows child elements to wrap, for example a link row followed by a search field."
+  )
+
   attr(:rest, :global,
     doc: """
     Additional HTML attributes added to the subnav element.
@@ -2369,6 +2374,7 @@ defmodule PrimerLive.Component do
     class =
       AttributeHelpers.classnames([
         "subnav",
+        assigns.is_wrap and "pl-subnav--wrap",
         assigns[:class]
       ])
 
@@ -3184,6 +3190,28 @@ defmodule PrimerLive.Component do
   </.text_input>
   ```
 
+  Add a button to the text input with slot `trailing_action`:
+
+  ```
+  <.text_input>
+    <:trailing_action>
+      <.button is_icon_only aria-label="Clear">
+        <.octicon name="x-16" />
+      </.button>
+    </:trailing_action>
+  </.text_input>
+  ```
+
+  Add a leading visual with slot `leading_visual`:
+
+  ```
+  <.text_input>
+    <:leading_visual>
+      <.octicon name="mail-16" />
+    </:leading_visual>
+  </.text_input>
+  ```
+
   Place the input inside a `form_group/1` with `is_form_group`. Attributes `form` and `field` are passed to the form group to generate a group label.
 
   ```
@@ -3268,7 +3296,8 @@ defmodule PrimerLive.Component do
       input: nil,
       input_group: nil,
       input_group_button: nil,
-      validation_message: nil
+      validation_message: nil,
+      input_wrap: nil
     },
     doc: """
     Additional classnames for input elements.
@@ -3282,6 +3311,7 @@ defmodule PrimerLive.Component do
       input_group: "",        # Wrapper around the grouped input and group button
       input_group_button: "", # Wrapper around slot group_button
       validation_message: "", # Validation message
+      input_wrap: "",         # Input wrapper when leading or trailing visual is used
     }
     ```
     """
@@ -3298,9 +3328,16 @@ defmodule PrimerLive.Component do
     doc: "Hide WebKit's contact info autofill icon."
   )
 
-  attr(:is_large, :boolean, default: false, doc: "Larger text size.")
+  attr(:is_large, :boolean,
+    default: false,
+    doc:
+      "Additional padding creates a higher input box. If no width is specified, the input box is slightly wider."
+  )
 
-  attr(:is_small, :boolean, default: false, doc: "Smaller input with smaller text size.")
+  attr(:is_small, :boolean,
+    default: false,
+    doc: "Smaller (less high) input with smaller text size."
+  )
 
   attr(:is_short, :boolean,
     default: false,
@@ -3310,6 +3347,16 @@ defmodule PrimerLive.Component do
   attr(:is_shorter, :boolean,
     default: false,
     doc: "Inside a form group. Generates an input with a even more reduced width."
+  )
+
+  attr(:is_monospace, :boolean,
+    default: false,
+    doc: "Uses a monospace font."
+  )
+
+  attr(:is_trailing_action_divider, :boolean,
+    default: false,
+    doc: "Adds a separator line to slot `trailing_action`."
   )
 
   attr(:is_form_group, :boolean,
@@ -3382,6 +3429,20 @@ defmodule PrimerLive.Component do
     """
   )
 
+  slot(:leading_visual,
+    required: false,
+    doc: """
+    Container for a leading visual. Commonly a `octicon/1` component is used.
+    """
+  )
+
+  slot(:trailing_action,
+    required: false,
+    doc: """
+    Container for a trailing action. Commonly a `octicon/1` component, or a `button/1` component with an icon (no label) is used.
+    """
+  )
+
   def text_input(assigns) do
     case validate_is_form(assigns) do
       {:error, reason} ->
@@ -3414,17 +3475,34 @@ defmodule PrimerLive.Component do
       valid?: valid?
     } = AttributeHelpers.common_input_attrs(assigns, nil)
 
+    type = assigns.type
+
+    has_leading_visual =
+      type !== "textarea" && !is_nil(assigns[:leading_visual]) && assigns[:leading_visual] !== []
+
+    has_trailing_action =
+      type !== "textarea" && !is_nil(assigns[:trailing_action]) &&
+        assigns[:trailing_action] !== []
+
+    has_input_wrap = has_leading_visual || has_trailing_action
+
     classes = %{
       input:
         AttributeHelpers.classnames([
-          "form-control",
-          assigns.is_contrast and "input-contrast",
+          if type === "textarea" do
+            "FormControl-textarea"
+          else
+            "FormControl-input"
+          end,
+          assigns.is_contrast and "FormControl-inset",
           assigns.is_hide_webkit_autofill and "input-hide-webkit-autofill",
-          assigns.is_large and "input-lg",
-          assigns.is_small and "input-sm",
-          assigns.is_short and "short",
-          assigns.is_shorter and "shorter",
-          assigns.is_full_width and "input-block",
+          !assigns.is_large and !assigns.is_small and "FormControl-medium",
+          assigns.is_large and "FormControl-large",
+          assigns.is_small and "FormControl-small",
+          assigns.is_full_width and "FormControl--fullWidth",
+          assigns.is_short and "FormControl--short",
+          assigns.is_shorter and "FormControl--shorter",
+          assigns.is_monospace and "FormControl-monospace",
           assigns.classes[:input],
           assigns.class
         ]),
@@ -3438,7 +3516,21 @@ defmodule PrimerLive.Component do
           "input-group-button",
           assigns.classes[:input_group_button]
         ]),
-      validation_message: assigns.classes[:validation_message]
+      validation_message: assigns.classes[:validation_message],
+      input_wrap:
+        AttributeHelpers.classnames([
+          "FormControl-input-wrap",
+          has_leading_visual and "FormControl-input-wrap--leadingVisual",
+          has_trailing_action and "FormControl-input-wrap--trailingAction",
+          assigns.classes[:input_wrap]
+        ]),
+      leading_visual: "FormControl-input-leadingVisualWrap",
+      trailing_action:
+        AttributeHelpers.classnames([
+          "FormControl-input-trailingAction",
+          type !== "textarea" && assigns.is_trailing_action_divider &&
+            "FormControl-input-trailingAction--divider"
+        ])
     }
 
     render = fn ->
@@ -3460,15 +3552,28 @@ defmodule PrimerLive.Component do
         )
 
       input =
-        apply(Phoenix.HTML.Form, FormHelpers.text_input_type_as_atom(assigns.type), [
+        apply(Phoenix.HTML.Form, FormHelpers.text_input_type_as_atom(type), [
           form,
           field,
           input_attrs
         ])
 
+      render_input_with_validation_marker = fn ->
+        assigns =
+          assigns
+          |> assign(:input, input)
+          |> assign(:validation_marker_attrs, validation_marker_attrs)
+
+        ~H"""
+        <%= if @validation_marker_attrs do %>
+          <span {@validation_marker_attrs}></span>
+        <% end %>
+        <%= @input %>
+        """
+      end
+
       assigns =
         assigns
-        |> assign(:input, input)
         |> assign(:phx_feedback_for_id, phx_feedback_for_id)
         |> assign(:classes, classes)
         |> assign(:has_group_button, has_group_button)
@@ -3477,24 +3582,31 @@ defmodule PrimerLive.Component do
         |> assign(:valid?, valid?)
         |> assign(:show_message?, show_message?)
         |> assign(:validation_message_class, classes.validation_message)
-        |> assign(:validation_marker_attrs, validation_marker_attrs)
+        |> assign(:has_input_wrap, has_input_wrap)
+        |> assign(:render_input_with_validation_marker, render_input_with_validation_marker)
 
       ~H"""
       <%= if @has_group_button do %>
         <div class={@classes.input_group}>
-          <%= if @validation_marker_attrs do %>
-            <span {@validation_marker_attrs}></span>
-          <% end %>
-          <%= @input %>
+          <%= @render_input_with_validation_marker.() %>
           <span class={@classes.input_group_button}>
             <%= render_slot(@group_button) %>
           </span>
         </div>
       <% else %>
-        <%= if @validation_marker_attrs do %>
-          <span {@validation_marker_attrs}></span>
+        <%= if @has_input_wrap do %>
+          <div class={@classes.input_wrap}>
+            <%= if !is_nil(@leading_visual) && @leading_visual !== [] do %>
+              <span class={@classes.leading_visual}><%= render_slot(@leading_visual) %></span>
+            <% end %>
+            <%= @render_input_with_validation_marker.() %>
+            <%= if !is_nil(@trailing_action) && @trailing_action !== [] do %>
+              <span class={@classes.trailing_action}><%= render_slot(@trailing_action) %></span>
+            <% end %>
+          </div>
+        <% else %>
+          <%= @render_input_with_validation_marker.() %>
         <% end %>
-        <%= @input %>
       <% end %>
       <.input_validation_message
         valid?={@valid?}
@@ -3652,6 +3764,7 @@ defmodule PrimerLive.Component do
 
   attr(:classes, :map,
     default: %{
+      select_container: nil,
       select: nil,
       validation_message: nil
     },
@@ -3663,6 +3776,7 @@ defmodule PrimerLive.Component do
     Default map:
     ```
     %{
+      select_container: "",   # Select container element
       select: "",             # Select element
       validation_message: "", # Validation message
     }
@@ -3681,11 +3795,20 @@ defmodule PrimerLive.Component do
     """
   )
 
-  attr :is_small, :boolean, default: false, doc: "Creates a small input."
+  attr :is_small, :boolean, default: false, doc: "Creates a small select."
+  attr :is_large, :boolean, default: false, doc: "Creates a large select."
+  attr :is_short, :boolean, default: false, doc: "Creates a short select."
+  attr :is_shorter, :boolean, default: false, doc: "Creates a shorter select."
+  attr :is_full_width, :boolean, default: false, doc: "Full width select."
 
   attr :is_auto_height, :boolean,
     default: false,
     doc: "When using `is_multiple`: sets the size to the number of options."
+
+  attr(:is_monospace, :boolean,
+    default: false,
+    doc: "Uses a monospace font."
+  )
 
   attr(:is_form_group, :boolean,
     default: false,
@@ -3746,13 +3869,27 @@ defmodule PrimerLive.Component do
       valid?: valid?
     } = AttributeHelpers.common_input_attrs(assigns, nil)
 
+    is_multiple = assigns.is_multiple
+
     classes = %{
+      select_container:
+        AttributeHelpers.classnames([
+          "FormControl-select-wrap",
+          is_multiple and "pl-multiple-select",
+          assigns.is_short and "FormControl--short",
+          assigns.is_shorter and "FormControl--shorter",
+          assigns.is_full_width and "FormControl--fullWidth",
+          assigns.classes[:select_container],
+          assigns[:class]
+        ]),
       select:
         AttributeHelpers.classnames([
-          "form-select",
-          assigns.is_small and "select-sm",
-          assigns.classes[:select],
-          assigns[:class]
+          "FormControl-select",
+          !assigns.is_large and !assigns.is_small and "FormControl-medium",
+          assigns.is_small and "FormControl-small",
+          assigns.is_large and "FormControl-large",
+          assigns.is_monospace and "FormControl-monospace",
+          assigns.classes[:select]
         ]),
       validation_message: assigns.classes[:validation_message]
     }
@@ -3760,7 +3897,8 @@ defmodule PrimerLive.Component do
     render = fn ->
       options = assigns.options
       is_auto_height = assigns.is_auto_height
-      is_multiple = assigns.is_multiple
+
+      container_attrs = [class: classes.select_container]
 
       input_attrs =
         AttributeHelpers.append_attributes(
@@ -3788,6 +3926,7 @@ defmodule PrimerLive.Component do
       assigns =
         assigns
         |> assign(:input, input)
+        |> assign(:container_attrs, container_attrs)
         |> assign(:phx_feedback_for_id, phx_feedback_for_id)
         |> assign(:classes, classes)
         |> assign(:validation_message_id, validation_message_id)
@@ -3798,10 +3937,12 @@ defmodule PrimerLive.Component do
         |> assign(:validation_marker_attrs, validation_marker_attrs)
 
       ~H"""
-      <%= if @validation_marker_attrs do %>
-        <span {@validation_marker_attrs}></span>
-      <% end %>
-      <%= @input %>
+      <div {@container_attrs}>
+        <%= if @validation_marker_attrs do %>
+          <span {@validation_marker_attrs}></span>
+        <% end %>
+        <%= @input %>
+      </div>
       <.input_validation_message
         valid?={@valid?}
         show_message?={@show_message?}
@@ -3893,7 +4034,7 @@ defmodule PrimerLive.Component do
     <:label>Some label</:label>
     <:disclosure>
       <span class="d-block mb-1">Available hours per week</span>
-      <input type="text" name="" class="form-control input-contrast" size="3" />
+      <.text_input is_contrast size="3" />
       <span class="text-small color-fg-muted pl-2">hours per week</span>
     </:disclosure>
   </.checkbox>